Also removing scooter_screen redundancy
